finalize KML export

+++ db/migrate/20130628130528_remove_projection_from_access_points.ninoxe_engine.rb
@@ -1,9 +1,15 @@
 # This migration comes from ninoxe_engine (originally 20130628124951)
 class RemoveProjectionFromAccessPoints < ActiveRecord::Migration
   def up
-    remove_column :access_points, :x
-    remove_column :access_points, :y
-    remove_column :access_points, :projection_type
+    if column_exists? :access_points, :x
+      remove_column :access_points, :x
+    end
+    if column_exists? :access_points, :y
+      remove_column :access_points, :y
+    end
+    if column_exists? :access_points, :projection_type
+      remove_column :access_points, :projection_type
+    end
   end
 
   def down
